requirements.txt
fastmcp>=2.1.2
pydantic>=2.0.0
docker>=6.0.0
python-dotenv>=1.0.0
typer>=0.15.0
rich>=14.0.0
pyyaml>=6.0.0
